page_store: rename pagefile => page_file
